requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
altair=4.2.2=pypi_0
attrs=23.2.0=pypi_0
backports-zoneinfo=0.2.1=pypi_0
blinker=1.8.1=pypi_0
ca-certificates=2024.3.11=hecd8cb5_0
cachetools=5.3.3=pypi_0
certifi=2024.2.2=py38hecd8cb5_0
charset-normalizer=3.3.2=pypi_0
click=8.1.7=pypi_0
coloredlogs=15.0.1=pypi_0
cycler=0.12.1=pypi_0
entrypoints=0.4=pypi_0
flask=3.0.3=pypi_0
flask-cors=4.0.0=pypi_0
flatbuffers=24.3.25=pypi_0
fonttools=4.51.0=pypi_0
gitdb=4.0.11=pypi_0
gitpython=3.1.43=pypi_0
humanfriendly=10.0=pypi_0
idna=3.7=pypi_0
imageio=2.9.0=pypi_0
importlib-metadata=7.1.0=pypi_0
importlib-resources=6.4.0=pypi_0
iniconfig=2.0.0=pypi_0
itsdangerous=2.2.0=pypi_0
jinja2=3.1.3=pypi_0
jsonschema=4.22.0=pypi_0
jsonschema-specifications=2023.12.1=pypi_0
kiwisolver=1.4.5=pypi_0
libcxx=14.0.6=h9765a3e_0
libedit=3.1.20230828=h6c40b1e_0
libffi=3.4.4=hecd8cb5_0
llvmlite=0.41.1=pypi_0
markdown-it-py=3.0.0=pypi_0
markupsafe=2.1.5=pypi_0
matplotlib=3.5.3=pypi_0
mdurl=0.1.2=pypi_0
more-itertools=10.2.0=pypi_0
mpmath=1.3.0=pypi_0
ncurses=6.4=hcec6c5f_0
networkx=3.1=pypi_0
numba=0.58.1=pypi_0
numpy=1.20.0=pypi_0
onnxruntime=1.16.3=pypi_0
opencv-python=4.9.0.80=pypi_0
opencv-python-headless=4.9.0.80=pypi_0
openssl=3.0.13=hca72f7f_0
packaging=24.0=pypi_0
pandas=1.0.5=pypi_0
pillow=8.1.2=pypi_0
pip=23.3.1=py38hecd8cb5_0
pkgutil-resolve-name=1.3.10=pypi_0
platformdirs=4.2.1=pypi_0
pluggy=0.13.1=pypi_0
pooch=1.8.1=pypi_0
prettytable=3.10.0=pypi_0
protobuf=3.20.3=pypi_0
py=1.11.0=pypi_0
pyarrow=16.0.0=pypi_0
pycocotools=2.0.7=pypi_0
pydeck=0.9.0=pypi_0
pyecharts=2.0.5=pypi_0
pygments=2.17.2=pypi_0
pymatting=1.1.12=pypi_0
pympler=1.0.1=pypi_0
pyparsing=3.1.2=pypi_0
pytest=6.0.1=pypi_0
python=3.8.19=h5ee71fb_0
python-dateutil=2.9.0.post0=pypi_0
pytz=2024.1=pypi_0
pywavelets=1.4.1=pypi_0
readline=8.2=hca72f7f_0
referencing=0.35.0=pypi_0
rembg=2.0.50=pypi_0
requests=2.31.0=pypi_0
rich=13.7.1=pypi_0
rpds-py=0.18.0=pypi_0
scikit-image=0.16.2=pypi_0
scipy=1.4.1=pypi_0
setuptools=68.2.2=py38hecd8cb5_0
simplejson=3.19.2=pypi_0
six=1.16.0=pypi_0
smmap=5.0.1=pypi_0
sqlite=3.41.2=h6c40b1e_0
streamlit=1.21.0=pypi_0
sympy=1.12=pypi_0
tabulate=0.9.0=pypi_0
tk=8.6.12=h5d9f67b_0
toml=0.10.2=pypi_0
toolz=0.12.1=pypi_0
tornado=6.4=pypi_0
tqdm=4.64.0=pypi_0
typing-extensions=4.11.0=pypi_0
tzdata=2024.1=pypi_0
tzlocal=5.2=pypi_0
urllib3=2.2.1=pypi_0
validators=0.28.1=pypi_0
wcwidth=0.2.13=pypi_0
werkzeug=3.0.2=pypi_0
wheel=0.37.1=pyhd3eb1b0_0
xz=5.4.6=h6c40b1e_0
zipp=3.18.1=pypi_0
zlib=1.2.13=h4dc903c_0
